Librarian Bot: Update Hugging Face dataset ID
This pull request updates the ID of the dataset used to train the model to the new Hub identifier `nyu-mll/glue` (which has been migrated moved from `glue`). We have been working to migrate datasets to their own repositories on the Hub, and this is part of that effort.

Updating the dataset ID in the model card will ensure that the model card is correctly linked to the dataset repository on the Hub. This will also make it easier for people to find your model via the training data used to create it.

This PR comes courtesy of [Librarian Bot](https://huggingface.co/librarian-bot). If you have any feedback, queries, or need assistance, please don't hesitate to reach out to [@davanstrien](https://huggingface.co/davanstrien).